Fix build failure on Win64 since r14458.
R=dslomov@chromium.org

Review URL: https://codereview.chromium.org/14150019

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14475 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/runtime.cc
@@ -2482,7 +2482,7 @@ RUNTIME_FUNCTION(MaybeObject*, Runtime_SuspendJSGeneratorObject) {
   // allocation fails.
   intptr_t offset = frame->pc() - function->code()->instruction_start();
   ASSERT(offset > 0 && Smi::IsValid(offset));
-  generator_object->set_continuation(offset);
+  generator_object->set_continuation(static_cast<int>(offset));
 
   // It's possible for the context to be other than the initial context even if
   // there is no stack handler active.  For example, this is the case in the
